refactor: Refactor kraken_api
